refactor(report): split gatekeeper teaser helpers to satisfy line guard
